fix: 공연 관련 정보 생성 api 수정
