+++ .gitignore
@@ -13,6 +13,7 @@ Thumbs.db
 *.ipr
 *.iws
 .vscode/
+.ai-context/
 
 # ---- Build tools ----
 .gradle/
